requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.1
bcp47==0.0.4
bioregistry==0.5.29
bleach==5.0.0
certifi==2021.10.8
CFGraph==0.2.1
chardet==4.0.0
charset-normalizer==2.0.12
class-resolver==0.3.10
click==8.1.3
commonmark==0.9.1
curies==0.1.5
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
distlib==0.3.4
docutils==0.17.1
et-xmlfile==1.1.0
fastobo==0.12.1
filelock==3.6.0
frozenlist==1.3.0
fsspec==2022.5.0
funowl==0.1.12
graphviz==0.20
greenlet==1.1.2
hbreader==0.9.1
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.3
isodate==0.6.1
Jinja2==3.1.2
joblib==1.1.0
json-flattener==0.1.9
jsonasobj==1.3.1
jsonasobj2==1.0.4
jsonpatch==1.32
jsonpath-ng==1.5.3
jsonpointer==2.3
jsonschema==4.4.0
keyring==23.6.0
kgcl==0.1.0
kgcl-rdflib==0.2.0
kgcl-schema==0.2.0
lark==1.1.2
linkml==1.2.14
linkml-dataops==0.1.0
linkml-runtime==1.2.16
markdown-it-py==2.1.0
MarkupSafe==2.1.1
mdit-py-plugins==0.3.0
mdurl==0.1.1
more-click==0.1.1
multidict==6.0.2
myst-parser==0.18.0
networkx==2.8
numpy==1.22.3
nxontology==0.4.1
oaklib==0.1.36
ontoportal-client==0.0.3
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pandasql==0.7.3
parse==1.19.0
pbr==5.8.1
pkginfo==1.8.3
platformdirs==2.5.2
ply==3.11
prefixcommons==0.1.9
pronto==2.5.0
pydantic==1.9.1
Pygments==2.12.0
PyJSG==0.11.10
pyparsing==2.4.7
pyrsistent==0.18.1
PyShEx==0.8.1
PyShExC==0.9.1
pystow==0.4.4
python-dateutil==2.8.2
PyTrie==0.4.0
pytz==2022.1
PyYAML==6.0
ratelimit==2.2.1
rdflib==6.1.1
rdflib-jsonld==0.6.1
rdflib-shim==1.0.3
readme-renderer==35.0
recommonmark==0.7.1
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rfc3987==1.3.8
rich==12.4.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
semsql==0.1.7
ShExJSG==0.8.2
six==1.16.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sparqlslurper==0.5.1
SPARQLWrapper==2.0.0
Sphinx==4.5.0
sphinx-click==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.36
SQLAlchemy-Utils==0.38.2
sssom==0.3.14
sssom-schema==0.9.4
stevedore==3.5.0
threadpoolctl==3.1.0
tqdm==4.64.0
twine==4.0.1
typing_extensions==4.2.0
urllib3==1.26.9
validators==0.18.2
virtualenv==20.14.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
watchdog==2.1.9
webencodings==0.5.1
wrapt==1.14.0
yarl==1.7.2
zipp==3.8.0
